make snowflake_connection have config info + stop using source_config/...
